+++ src/navigation/ShoppingLists/ShoppingListsRouteParams.ts
@@ -1,12 +1,19 @@
 import List from '../../models/List'
+import Product from '../../models/Product'
+import Category from '../../models/Category'
 
 type ShoppingListsRouteParams = {
     ShoppingListsScreen: undefined
     ShoppingListDetailsScreen: { list: List }
     NewListScreen: undefined
-    AddProductsScreen: undefined
+    AddProductsScreen: { list: List }
     ProductDetailsScreen: undefined
-    CategoriesListScreen: undefined
+    CategoriesListScreen: {
+        product?: Product
+        list?: List
+        setCategory?: (category: Category) => void
+        chosenCategory?: Category
+    }
 }
 
 export default ShoppingListsRouteParams